Move text face/color parsing functions into new module
No other change.
